chore: remove old root command
